Changing Arm command ID's. Adding documentation to PwmReader
RED currently uses 2015's command ID's for the arm. So I coded up 2015's
commands as well as 2016's, and now the program can recognize both
versions.

As for PwmReader, I added a lot of documentation and fixed a couple of
remaining issues, such as the reader not knowing what the pin state was
on startup and forgetting to reset the 'edge recieved' variable along
with everything else when 0% duty cycle is detected