PWM reader works when not 0/100% duty, 0/100% conditions still buggy
I tested out with an oscilloscope, and when not at 100% or 0% duty cycle
the program managed to correctly return duty, period on, and period off
within only a small margin of error of +- a few microseconds.

The code for handling 0% and 100%, is currently disabled as it was quite
buggy, I'll need to get it working later. Documentation is somewhat out
of date, due to this; the functions and descriptions all related to non
0% and 100% remain accurate.