Replaced ControlFrameworkInterface with JointInterface with 3 smaller classes
Added SingleMotorJoint, TiltJoint and RotateJoint classes to the
framework allowing for multiple motor control on a single joint.
Documented the new classes. JointInterface is the super class for the
three new classes containing all shared components of the 3 types.
Selection of algorithm was made into a function in JointInterface.
Seperate constructors for use with feedback and without feedback.