Made minor logic change involving the orientation of the Motors on a Two motor joint
The JointInterface sees the arm move in the same direction. The
controllers move in opposite directions since they are opposite one
another. The interface doesnt know this so the controller alone makes
this change. For rotation they move opposite as far as the controller is
concerned so they are sent opposite signals and then convert them as
need be.