PWM reader functioning on all pins, tested at 4096 us pwm
I tested it, it seemed to spit out the correct on period and off period
times on all the noted pins. I editted the header to include a list of
pins you can use. 0% and 100% works, though it takes the program approx
1 second to recongize it

Needs fixes to documentation, possibly refactoring at some point as it
got a little ugly, and needs all five tested at once