adding shitty module for last year's shit motor controller, SDC2130
Added module for it. The motor controller can move the motor based on
speed position and torque, and be talked to with pwm and serial. But
only pwm is implemented, and due to it being a shitty motor controller
the only way we can currently reliably move it is to increment it
slightly one way or another, and that's when you pass the module a speed
value.

Did that because that's how they did it last year, which is shitty but
we're unlikely to use this motor controller again anyway so it mainly
exists for reverse arm compatibality