Fixing bug where DynaController move output bad values. More documentation
DynaController's move function would output the movement variable, not
the send variable. Movement variable was the unmodified value first
passed into the function, while send was the variable that had been
modified for actual usage. It needs to be the one used in the
dynamixelSpinWheel function.

More documentation and fixed some spacing