Closed loop algorithms now have to be passed into interface constructor
Closed loop algorithms are generally too complex to just generalize
away, and will require main program modification for special instances.
So, the framework now requires the user to pass in an IOAlgorithm
instance whenever it's calling the joint interface constructor that
takes in a feedback device.

For the constructor that doesn't take in feedback devices, IE for use if
the joint has open loop control, the old procedure remains unchanged.
That is, the user just passes in an input type and an output device, and
internally the joint interface class will select a dumb algorithm for
the user