Joint interface commands now return enum, represents output states
The joint interface commands used by the main program now return an enum
that represents what happened when the joint interface attempted to make
the user's desired output happen. For instance, it can return states
that say the user passed an invalid input, the interface wasn't properly
constructed, that the output was ran but the joint needs more control
loop calls to finish getting to its desired output, or if the output was
ran and the joint is now at the desired state.

Also restructed the folders a bit, is why there looks like there's so
many deletes