Correcting bug with speed-to-pwm mapping
we used SPD_MIN to map, but problem is we'd basically be mapping
negative max magnitude to pos max magnitude, to 0 pwm magnitude to max
pwm magnintude. Two different scales.

So I set it so it mapped 0 to speed max instead, which is the scale it's
supposed to map to